external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/glsl/misc/shader-without-precision.frag-expected.txt

CONSOLE MESSAGE: line 24: *** Error compiling shader '[object WebGLShader]':ERROR: 0:2: '' : No precision specified for (float) 

fragment shader without precision should fail

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test: fragment shader without precision should fail
default vertex shader
test fragment shaderPASS fragment shader without precision should fail
PASS successfullyParsed is true

TEST COMPLETE
